\name{Y}
\alias{Y}
\docType{data}
\title{
Observation matrix Y
}
\description{
An example of observation matrix
}
\usage{data("Y")}
\format{
  The format is:
 num [1:300, 1:15] 3 1 1 0 0 3 2 0 3 2 ...
}
\references{
M. Gomtsyan et al. "Variable selection in sparse multivariate GLARMA models: Application to germination control by environment", arXiv:2208.14721
}
\examples{
data(Y)
}
